Adding benchmarks for the Stiletto IoC container.
Stiletto doesn't have a method 'object Container.Resolve(object)';
instead, it has only 'T Container.Get<T>()'.  Because of this,
ContainerAdapterBase was modified to add a virtual method of the same
signature, whose default implementation is just a call to the old
'Resolve(object)' abstract method.  Other than that, this is as
straightforward an addition as could be hoped.